добавлен файл с созданием таблиц,в боте добавлены дополнительные проверки на корректность введеных ссылок

+++ Adds_Links_Bot.py
@@ -7,7 +7,6 @@ from aiogram.contrib.fsm_storage.memory import MemoryStorage
 from aiogram.dispatcher import FSMContext
 from aiogram.dispatcher.filters.state import State, StatesGroup
 from dotenv import load_dotenv
-
 load_dotenv()
 
 BOT_TOKEN = os.getenv("BOT_TOKEN")
@@ -46,29 +45,36 @@ async def download_command(message: types.Message):
 async def download_links(message: types.Message, state: FSMContext):
     async with aiohttp.ClientSession() as session:
         urls = re.findall(r'(?:https?://\S+)', message.text)
-        file_path = "chats_users.xlsx"
-        chat_ids = []
-        for url in urls:
-            async with session.get(f"http://{IP_USERS_SAVE}/chats_links?urls={url}") as resp:
-                try:
-                    if resp.status == 200:
-                        file_content = await resp.read()
-                        with open(file_path, "ab") as f:
-                            f.write(file_content)
-                    else:
-                        error_message = await resp.text()
-                        if resp.status == 404:
-                            chat_ids.append(url)
-                except Exception as e:
-                    await message.reply(f"Произошла ошибка при обработке запроса: {e}")
-                finally:
-                    await session.close()
-        if chat_ids:
-            await message.reply(f"{error_message}\n" + '\n'.join(chat_ids))
-        if os.path.exists(file_path):
-            with open(file_path, "rb") as f:
-                await message.reply_document(f)
-                os.remove(file_path)
+        if urls:
+            file_path = "chats_users.xlsx"
+            invalid_chat_ids_server = []
+            invalid_chat_ids = []
+            for url in urls:
+                remainder = url.split("://")[1].split("/", 1)[-1]
+                if re.match(r'^https?://', url) and not '/' in remainder:
+                    async with session.get(f"http://{IP_USERS_SAVE}/chats_links?urls={url}") as resp:
+                        try:
+                            if resp.status == 200:
+                                file_content = await resp.read()
+                                with open(file_path, "ab") as f:
+                                    f.write(file_content)
+                            else:
+                                error_message = await resp.text()
+                                invalid_chat_ids_server.append(url)
+                        except Exception as e:
+                            await message.reply(f"Произошла ошибка при обработке запроса: {e}")
+                else:
+                    invalid_chat_ids.append(url)
+            if invalid_chat_ids_server:
+                await message.reply(f"{error_message}\n" + '\n'.join(invalid_chat_ids_server))
+            if invalid_chat_ids:
+                await message.reply(f'Ссылки должны начинаться с "https://" и не содержать "/" в конце. Например, ссылка "https://t.me/example1/4544" неправильна, так как содержит "/4544" в конце.\n' +'\n'.join(invalid_chat_ids_server))
+            if os.path.exists(file_path):
+                with open(file_path, "rb") as f:
+                    await message.reply_document(f)
+                    os.remove(file_path)
+        else:
+            await message.reply(f'Ссылки должны начинаться с "https://" и не содержать "/" в конце. Например, ссылка "https://t.me/example1/4544" неправильна, так как содержит "/4544" в конце.')
         await state.finish()
 
 
@@ -76,24 +82,26 @@ async def download_links(message: types.Message, state: FSMContext):
 @dp.message_handler(state=ParseState.waiting_for_tasks_links)
 async def tasks_links(message: types.Message, state: FSMContext):
     links = re.findall(r'(?:https?://\S+)', message.text)
-    valid_links = []
-    invalid_links = []
+    if links:
+        valid_links = []
+        invalid_links = []
 
-    for link in links:
-        remainder = link.split("://")[1].split("/", 1)[-1]
-        if re.match(r'^https?://', link) and not '/' in remainder:
-            valid_links.append(link)
+        for link in links:
+            remainder = link.split("://")[1].split("/", 1)[-1]
+            if re.match(r'^https?://', link) and not '/' in remainder:
+                valid_links.append(link)
+            else:
+                invalid_links.append(link)
+
+        if invalid_links:
+            await message.reply('Ссылки должны начинаться с "https://" и не содержать "/" в конце. Например, ссылка "https://t.me/example1/4544" неправильна, так как содержит "/4544" в конце.\n' + '\n'.join(invalid_links))
         else:
-            invalid_links.append(link)
-
-    if invalid_links:
-        await message.reply('Обнаружены недопустимые ссылки:\n' + '\n'.join(invalid_links))
+            if valid_links:
+                requests.get(f"http://{IP}/add")
+                requests.post(f"http://{IP}/add", json={"urls": valid_links})
+                await message.reply('Чаты успешно добавлены в очередь для парсинга.Рекомендуется подождать час, перед тем как получить информацию о пользователях из чатов.')
     else:
-        if valid_links:
-            requests.get(f"http://{IP}/add")
-            requests.post(f"http://{IP}/add", json={"urls": valid_links})
-            await message.reply('Ссылки добавлены')
-
+        await message.reply(f'Ссылки должны начинаться с "https://" и не содержать "/" в конце. Например, ссылка "https://t.me/example1/4544" неправильна, так как содержит "/4544" в конце.')
     await state.finish()
 
 async def on_shutdown():
